postinst: better DESTDIR, always re-process all apps, fix direct links.
Of note, the switch to re-processing all apps. Not doing so creates a
strong tie between package names and names of application profiles.
For instance, for apt, we must use "apt" in the profiles. That sounds fine
but consider gnutls which produces several binary packages, which could
change, and also consider its source package name: gnutls28. That's too
user-unfriendly and the script currently only takes a fraction of a second
(50ms).