Use noawait in triggers where possible
Per deb-triggers (5): The "noawait" variant should be used when the
functionality provided by the trigger is not crucial.

I'm not 100% certain about this change: it rather depends on whether you
want a failure in crypto-config's run to prevent the configuration of
the package that triggered it (in the Debian packaging sense). You may
want to revert this, but I would recommend testing the "await" variant
to make sure it does what you expect.